+++ tests/test_rice.py
@@ -8,8 +8,8 @@
 import pytest
 from abses import MainModel, PatchModule
 
-from abses_sce.env import CompetingCell
-from abses_sce.rice_farmer import RiceFarmer
+from src.env import CompetingCell
+from src.rice_farmer import RiceFarmer
 
 from .conftest import cfg, set_cell_arable_condition
